+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
 - Fixed missing misc. button and link type definition exports ([#2434](https://github.com/elastic/eui/pull/2434))
 - Strip custom semantics from `EuiSideNav` ([#2429](https://github.com/elastic/eui/pull/2429))
+- Changed `EuiLink` to appear uninteractive when given the disabled prop and an onClick handler ([#2423](https://github.com/elastic/eui/pull/2423))
+
 
 ## [`14.5.0`](https://github.com/elastic/eui/tree/v14.5.0)
 
@@ -32,6 +34,7 @@
 - Added `keyboardShorcut` glyph to 'EuiIcon ([#2413](https://github.com/elastic/eui/pull/2413))
 - Improved a11y in `EuiNavDrawer` ([#2417](https://github.com/elastic/eui/pull/2417))
 - Improved a11y in `EuiSuperDatePicker` ([#2426](https://github.com/elastic/eui/pull/2426))
+
 
 **Bug fixes**